requirements.txt
aniso8601==9.0.1
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
dnspython==2.3.0
elastic-transport==8.4.0
elasticsearch==7.17.9
Flask==2.2.3
Flask-JWT-Extended==4.4.4
flask-restx==1.1.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
pkgutil-resolve-name==1.3.10
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
requests==2.28.2
six==1.16.0
tomli==2.0.1
Unidecode==1.3.6
urllib3==1.26.15
uuid==1.30
Werkzeug==2.2.3
xmltodict==0.13.0
zipp==3.15.0
